Multiple bug fixes.
ASABASE-38 - Device name is being reset to an empty string on server sample of java
ASABASE-36 - Changing the DefaultLanguage to unsupported language gives the wrong error message.
Improved the .classpath for the config service.  Added usage of the about.jar
ASABASE-127 Remove the addMatch method call from the application layer of the sample applications

Change-Id: If33464aa29a05cf894a63b6388d3181296b972a3
Signed-off-by: irivin <irivin@qce.qualcomm.com>
Signed-off-by: Ry Jones <c_lawren@qce.qualcomm.com>