ASACORE-1984 Add the -Wunused-parameter flag to Linux build
The Windows build fails when there is an unused parameter, however
the Linux build does not fail when there is an unused parameter.
This makes it difficult to know the build will fail when working
on Linux. Adding the -Wunused-parameter flag to the Linux build will
cause the Linux to fail when there is an unused parameter similar to
windows. Making the build warnings similar across the OSs will help
reduce developer issues when committing code for the verification
build.

Change-Id: I9510da3504cc4dda15876b4dff9549e835e63fdf
Signed-off-by: George Nash <georgen@qce.qualcomm.com>