ASABASE-579 and ASABASE-580: NotificationService andoid sample failure
Fixed bug where the NotificationServiceUISample could fail to
set the sample as producer and/or consumer. This was happening
because the AboutObj var was being set to null during the lifetime
of its corresponding bus attachment. Once the AboutObj is created
it must remain for the lifetime of the bus attachment; instead,
the AboutObj can simply be unannounced and its port unbounded.
Only when the bus attachment is released is it ok to then set
the AboutObj variable to null.

Change-Id: I88956bea1a549b2a2c474a8f9ca413b8f2b1fa90
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>